added pullfile write guard to cronjob

+++ capito/haleres/cron.py
@@ -57,15 +57,16 @@ if ipc_folder_list:
     pullfile.write_text("\n".join(ipc_folder_list))
     # Call rsync with pullfile - intentionally blocking!
     log_list.append(f"Pulling {len(ipc_folder_list)} ipc-folder(s). Pull-File: {str(pullfile)}")
-    subprocess.check_output([
-        "rsync", 
-        "-ar",
-        "--ignore-missing-args",
-        f"--files-from={str(pullfile)}",
-        f"{hlrs_server}:{haleres_settings.workspace_path}/",
-        f"{haleres_settings.mount_point}"
-    ])
-    pullfile.unlink()
+    if pullfile.exists():
+        subprocess.check_output([
+            "rsync", 
+            "-ar",
+            "--ignore-missing-args",
+            f"--files-from={str(pullfile)}",
+            f"{hlrs_server}:{haleres_settings.workspace_path}/",
+            f"{haleres_settings.mount_point}"
+        ])
+        pullfile.unlink()
 
 # PULL IMAGES
 # Create pull list
